meek rules integration in ContinuousPC and ContinuousMIIC
