+++ lib/src/CMakeLists.txt
@@ -12,9 +12,7 @@ ot_add_source_file ( IndicesManip.cxx )
 ot_add_source_file ( ContinuousBayesianNetwork.cxx )
 ot_add_source_file ( ContinuousBayesianNetworkFactory.cxx )
 ot_add_source_file ( ContinuousPC.cxx )
-ot_add_source_file ( ContinuousPC2.cxx )
 ot_add_source_file ( ContinuousMIIC.cxx )
-ot_add_source_file ( ContinuousMIIC2.cxx )
 ot_add_source_file ( Greater.cxx )
 ot_add_source_file ( TabuList.cxx )
 ot_add_source_file ( JunctionTreeBernsteinCopula.cxx )
@@ -31,9 +29,7 @@ ot_install_header_file ( IndicesManip.hxx )
 ot_install_header_file ( ContinuousBayesianNetwork.hxx )
 ot_install_header_file ( ContinuousBayesianNetworkFactory.hxx )
 ot_install_header_file ( ContinuousPC.hxx )
-ot_install_header_file ( ContinuousPC2.hxx )
 ot_install_header_file ( ContinuousMIIC.hxx )
-ot_install_header_file ( ContinuousMIIC2.hxx )
 ot_install_header_file ( Greater.hxx )
 ot_install_header_file ( TabuList.hxx )
 ot_install_header_file ( JunctionTreeBernsteinCopula.hxx )
