+++ lib/test/CMakeLists.txt
@@ -41,10 +41,8 @@ ot_check_test ( IndicesManip_std )
 ot_check_test ( StratifiedCache_std )
 ot_check_test ( ContinuousTTest_std )
 ot_check_test ( ContinuousPC_std )
-ot_check_test ( ContinuousPC2_std )
 ot_check_test ( CorrectedMutualInformation_std )
 ot_check_test ( ContinuousMIIC_std )
-ot_check_test ( ContinuousMIIC2_std )
 ot_check_test ( TabuList_std )
 ot_check_test ( ContinuousBayesianNetwork_std )
 ot_check_test ( OTAgrumResourceMap_std )
