--- lib/test/t_ContinuousMIIC2_std.expout
@@ -1,146 +1,0 @@
-Sample size: 4000
-Sample dimension:7
-Sample 
-       [ A1         A2         B1         ... X0         X1         X2         ]
-   0 : [ 0.629877   0.887754   0.148519   ... 0.0296484  0.0245951  0.00461582 ]
-   1 : [ 0.135276   0.013763   0.706989   ... 0.59399    0.486341   0.49929    ]
-   2 : [ 0.347057   0.912595   0.846531   ... 0.449886   0.71616    0.888728   ]
-...
-3997 : [ 0.544293   0.658496   0.127842   ... 0.802371   0.53486    0.695887   ]
-3998 : [ 0.341802   0.160431   0.636882   ... 0.987952   0.971837   0.929303   ]
-3999 : [ 0.794498   0.689409   0.434064   ... 0.618516   0.8981     0.874322   ]
-
-
-go
-
-****
-Skeleton
-****
-digraph "no_name" {
-edge [dir=none]
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 1;
-  2 -> 3;
-  4 -> 6;
-  4 -> 5;
-  5 -> 6;
-
-}
-
-
-****
-PDAG
-****
-digraph "no_name" {
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 1 [dir=none];
-  2 -> 3 [dir=none];
-  4 -> 6 [dir=none];
-  4 -> 5 [dir=none];
-  5 -> 6 [dir=none];
-
-}
-
-
-****
-Named DAG
-****
-[A1,A2,B1,B2,X0,X1,X2]
-[X1->X2,X1->X0,X2->X0,B2->B1,A2->A1]
-
-digraph {
-    "X1"->"X2"
-    "X1"->"X0"
-    "X2"->"X0"
-    "X0"
-    "B2"->"B1"
-    "B1"
-    "A2"->"A1"
-    "A1"
-}
-
-[x1,x2,x3,x4,x5,x6,x7,y11,y12,y13,y21,y22,y31]#13
-PDAG: {0,1,2,3,4,5,6,7,8,9,10,11,12} , {3->11,7->11,0->9,6->9,1->7,3->8,12->10,11->12,0->11,5->9,2->7,10->8,2->12,10->4} , {}
-DAG: [x1,x2,x3,x4,x5,x6,x7,y11,y12,y13,y21,y22,y31]
-[x7->y13,x6->y13,x4->y22,x4->y12,x3->y11,x3->y31,x2->y11,y11->y22,x1->y22,x1->y13,y22->y31,y31->y21,y21->y12,y21->x5]
-
-Number of nodes: 7
-Sample size : 5000
-Sample dimension : 7
-Description : [A,B,C,D,E,F,G]
-
-
-go
-
-****
-skeleton
-****
-digraph "no_name" {
-edge [dir=none]
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 1;
-  0 -> 4;
-  0 -> 2;
-  0 -> 5;
-  1 -> 3;
-  1 -> 2;
-  5 -> 6;
-
-}
-
-
-****
-PDAG
-****
-digraph "no_name" {
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 4;
-  0 -> 2;
-  1 -> 3 [dir=none];
-  1 -> 2 [dir=none];
-  1 -> 0;
-  5 -> 6 [dir=none];
-  5 -> 0;
-
-}
-
-
-****
-CPDAG
-****
-digraph "no_name" {
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 4;
-  0 -> 2;
-  1 -> 3 [dir=none];
-  1 -> 0;
-  1 -> 2;
-  5 -> 6 [dir=none];
-  5 -> 0;
-
-}
-
-
-****
-Named DAG
-****
-[A,B,C,D,E,F,G]
-[G->F,F->A,D->B,B->A,B->C,A->E,A->C]
-
-digraph {
-    "G"->"F"
-    "F"->"A"
-    "D"->"B"
-    "B"->"A"
-    "B"->"C"
-    "A"->"E"
-    "A"->"C"
-    "C"
-    "E"
-}
-
