+++ lib/test/t_ContinuousMIIC_std.expout
@@ -30,7 +30,7 @@ node [shape = ellipse];
 
 
 ****
-CPDAG
+PDAG
 ****
 digraph "no_name" {
 node [shape = ellipse];
@@ -48,13 +48,13 @@ node [shape = ellipse];
 Named DAG
 ****
 [A1,A2,B1,B2,X0,X1,X2]
-[X2->X0,X2->X1,X0->X1,B2->B1,A2->A1]
+[X1->X2,X1->X0,X2->X0,B2->B1,A2->A1]
 
 digraph {
+    "X1"->"X2"
+    "X1"->"X0"
     "X2"->"X0"
-    "X2"->"X1"
-    "X0"->"X1"
-    "X1"
+    "X0"
     "B2"->"B1"
     "B1"
     "A2"->"A1"
@@ -65,4 +65,65 @@ digraph {
 PDAG: {0,1,2,3,4,5,6,7,8,9,10,11,12} , {3->11,7->11,0->9,6->9,1->7,3->8,12->10,11->12,0->11,5->9,2->7,10->8,2->12,10->4} , {}
 DAG: [x1,x2,x3,x4,x5,x6,x7,y11,y12,y13,y21,y22,y31]
 [x7->y13,x6->y13,x4->y22,x4->y12,x3->y11,x3->y31,x2->y11,y11->y22,x1->y22,x1->y13,y22->y31,y31->y21,y21->y12,y21->x5]
+
+Number of nodes: 7
+Sample size : 5000
+Sample dimension : 7
+Description : [A,B,C,D,E,F,G]
+
+
+go
+
+****
+skeleton
+****
+digraph "no_name" {
+edge [dir=none]
+node [shape = ellipse];
+  0;  1;  2;  3;  4;  5;  6;
+  0 -> 1;
+  0 -> 4;
+  0 -> 2;
+  0 -> 5;
+  1 -> 3;
+  1 -> 2;
+  5 -> 6;
+
+}
+
+
+****
+PDAG
+****
+digraph "no_name" {
+node [shape = ellipse];
+  0;  1;  2;  3;  4;  5;  6;
+  0 -> 4;
+  0 -> 2;
+  1 -> 3 [dir=none];
+  1 -> 2 [dir=none];
+  1 -> 0;
+  5 -> 6 [dir=none];
+  5 -> 0;
+
+}
+
+
+****
+Named DAG
+****
+[A,B,C,D,E,F,G]
+[G->F,F->A,D->B,B->A,B->C,A->E,A->C]
+
+digraph {
+    "G"->"F"
+    "F"->"A"
+    "D"->"B"
+    "B"->"A"
+    "B"->"C"
+    "A"->"E"
+    "A"->"C"
+    "C"
+    "E"
+}
