--- lib/test/t_ContinuousPC2_std.expout
@@ -1,218 +1,0 @@
-Sample size : 5000
-Sample dimension :7
-Sample        [ A1        A2        B1        ... X0        X1        X2        ]
-   0 : [ 0.629877  0.887754  0.424722  ... 0.721616  0.99309   0.931811  ]
-   1 : [ 0.135276  0.013763  0.478569  ... 0.50062   0.175258  0.400078  ]
-   2 : [ 0.347057  0.912595  0.700097  ... 0.0542298 0.111615  0.01064   ]
-...
-4997 : [ 0.780708  0.861619  0.57336   ... 0.814699  0.840734  0.936302  ]
-4998 : [ 0.0667799 0.223736  0.7       ... 0.538345  0.0951127 0.546137  ]
-4999 : [ 0.211817  0.0681641 0.461658  ... 0.810161  0.395742  0.263888  ]
-go
-digraph "skeleton" {
-  edge [dir = none];
-  node [shape = ellipse];
-  0; 1; 2; 3; 4; 5; 6; 
-  0->1 [label="t=67.3
-p=0"]
-  4->5 [label="t=61.1
-p=1.02e-20"]
-  4->6 [label="t=64.9
-p=1.56e-22"]
-  2->3 [label="t=122
-p=0"]
-  5->6 [label="t=58
-p=5.29e-16"]
-}
-digraph "no_name" {
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 1 [dir=none];
-  2 -> 3 [dir=none];
-  4 -> 6 [dir=none];
-  4 -> 5 [dir=none];
-  5 -> 6 [dir=none];
-
-}
-
-0--1 : p-value=0   ttest=67.3106
-4--5 : p-value=1.02025e-20   ttest=61.0624
-4--6 : p-value=1.56447e-22   ttest=64.8518
-2--3 : p-value=0   ttest=121.91
-5--6 : p-value=5.28577e-16   ttest=58.0097
-000 :      A1-X1|                      p=0.991163    ttest=-0.0110752
-001 :      A2-B2|                      p=0.88989     ttest=-0.138443
-002 :      B1-X1|                      p=0.969042    ttest=-0.0388093
-003 :      B2-X0|                      p=0.823654    ttest=0.222848
-004 :      A2-X0|                      p=0.81224     ttest=-0.237537
-005 :      A1-B1|                      p=0.973887    ttest=-0.0327334
-006 :      A1-X2|                      p=0.86512     ttest=-0.16986
-007 :      B1-X2|                      p=0.866631    ttest=-0.16794
-008 :      A2-X1|                      p=0.976229    ttest=-0.0297972
-009 :      A1-B2|                      p=0.997611    ttest=-0.00299442
-010 :      B2-X1|                      p=0.92925     ttest=-0.088788
-011 :      B1-X0|                      p=0.917316    ttest=-0.103815
-012 :      A1-X0|                      p=0.824288    ttest=0.222034
-013 :      A2-B1|                      p=0.947551    ttest=-0.0657827
-014 :      A2-X2|                      p=0.941704    ttest=-0.0731289
-015 :      B2-X2|                      p=0.996491    ttest=-0.00439842
-
-
-go
-
-****
-skeleton
-****
-digraph "no_name" {
-edge [dir=none]
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 1;
-  2 -> 3;
-  4 -> 6;
-  4 -> 5;
-  5 -> 6;
-
-}
-
-
-****
-PDAG
-****
-digraph "no_name" {
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 1 [dir=none];
-  2 -> 3 [dir=none];
-  4 -> 6 [dir=none];
-  4 -> 5 [dir=none];
-  5 -> 6 [dir=none];
-
-}
-
-
-****
-Named Junction tree
-****
-[A1,A2,B1,B2,X0,X1,X2]
-0 : [3(B2),2(B1)]
-1 : [6(X2),4(X0),5(X1)]
-2 : [1(A2),0(A1)]
-
-
-****
-Named DAG
-****
-[A1,A2,B1,B2,X0,X1,X2]
-[X1->X2,X1->X0,X2->X0,B2->B1,A2->A1]
-
-digraph {
-    "X1"->"X2"
-    "X1"->"X0"
-    "X2"->"X0"
-    "X0"
-    "B2"->"B1"
-    "B1"
-    "A2"->"A1"
-    "A1"
-}
-
-000 :      A1-X1|                      p=0.991163    ttest=-0.0110752
-001 :      A2-B2|                      p=0.88989     ttest=-0.138443
-002 :      B1-X1|                      p=0.969042    ttest=-0.0388093
-003 :      B2-X0|                      p=0.823654    ttest=0.222848
-004 :      A2-X0|                      p=0.81224     ttest=-0.237537
-005 :      A1-B1|                      p=0.973887    ttest=-0.0327334
-006 :      A1-X2|                      p=0.86512     ttest=-0.16986
-007 :      B1-X2|                      p=0.866631    ttest=-0.16794
-008 :      A2-X1|                      p=0.976229    ttest=-0.0297972
-009 :      A1-B2|                      p=0.997611    ttest=-0.00299442
-010 :      B2-X1|                      p=0.92925     ttest=-0.088788
-011 :      B1-X0|                      p=0.917316    ttest=-0.103815
-012 :      A1-X0|                      p=0.824288    ttest=0.222034
-013 :      A2-B1|                      p=0.947551    ttest=-0.0657827
-014 :      A2-X2|                      p=0.941704    ttest=-0.0731289
-015 :      B2-X2|                      p=0.996491    ttest=-0.00439842
-[x1,x2,x3,x4,x5,x6,x7,y11,y12,y13,y21,y22,y31]#13
-pdag:{0,1,2,3,4,5,6,7,8,9,10,11,12} , {6->9,0->9,5->9} , {7--11,2--7,8--10,11--12}
-dag:[x1,x2,x3,x4,x5,x6,x7,y11,y12,y13,y21,y22,y31]
-[y31->y22,y22->y11,y11->x3,y21->y12,x7->y13,x6->y13,x1->y13]
-
-Number of nodes: 7
-Sample size : 5000
-Sample dimension : 7
-Description : [A,B,C,D,E,F,G]
-
-
-go
-
-****
-skeleton
-****
-digraph "no_name" {
-edge [dir=none]
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 1;
-  0 -> 4;
-  0 -> 2;
-  0 -> 5;
-  1 -> 3;
-  1 -> 2;
-  5 -> 6;
-
-}
-
-
-****
-PDAG
-****
-digraph "no_name" {
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 4 [dir=none];
-  0 -> 2 [dir=none];
-  1 -> 3 [dir=none];
-  1 -> 2 [dir=none];
-  1 -> 0;
-  5 -> 6 [dir=none];
-  5 -> 0;
-
-}
-
-
-****
-CPDAG
-****
-digraph "no_name" {
-node [shape = ellipse];
-  0;  1;  2;  3;  4;  5;  6;
-  0 -> 4;
-  0 -> 2;
-  1 -> 3 [dir=none];
-  1 -> 0;
-  1 -> 2;
-  5 -> 6 [dir=none];
-  5 -> 0;
-
-}
-
-
-****
-Named DAG
-****
-[A,B,C,D,E,F,G]
-[G->F,F->A,D->B,B->A,B->C,A->E,A->C]
-
-digraph {
-    "G"->"F"
-    "F"->"A"
-    "D"->"B"
-    "B"->"A"
-    "B"->"C"
-    "A"->"E"
-    "A"->"C"
-    "C"
-    "E"
-}
-
