+++ lib/test/t_ContinuousPC_std.expout
@@ -77,7 +77,7 @@ node [shape = ellipse];
 
 
 ****
-CPDAG
+PDAG
 ****
 digraph "no_name" {
 node [shape = ellipse];
@@ -104,17 +104,17 @@ Named Junction tree
 Named DAG
 ****
 [A1,A2,B1,B2,X0,X1,X2]
-[X0->X2,X0->X1,X2->X1,B1->B2,A1->A2]
+[X1->X2,X1->X0,X2->X0,B2->B1,A2->A1]
 
 digraph {
-    "X0"->"X2"
-    "X0"->"X1"
-    "X2"->"X1"
-    "X1"
-    "B1"->"B2"
-    "B2"
-    "A1"->"A2"
-    "A2"
+    "X1"->"X2"
+    "X1"->"X0"
+    "X2"->"X0"
+    "X0"
+    "B2"->"B1"
+    "B1"
+    "A2"->"A1"
+    "A1"
 }
 
 000 :      A1-X1|                      p=0.991163    ttest=-0.0110752
@@ -136,5 +136,66 @@ digraph {
 [x1,x2,x3,x4,x5,x6,x7,y11,y12,y13,y21,y22,y31]#13
 pdag:{0,1,2,3,4,5,6,7,8,9,10,11,12} , {6->9,0->9,5->9} , {7--11,2--7,8--10,11--12}
 dag:[x1,x2,x3,x4,x5,x6,x7,y11,y12,y13,y21,y22,y31]
-[y12->y21,x7->y13,x6->y13,x3->y11,y11->y22,y22->y31,x1->y13]
+[y31->y22,y22->y11,y11->x3,y21->y12,x7->y13,x6->y13,x1->y13]
+
+Number of nodes: 7
+Sample size : 5000
+Sample dimension : 7
+Description : [A,B,C,D,E,F,G]
+
+
+go
+
+****
+skeleton
+****
+digraph "no_name" {
+edge [dir=none]
+node [shape = ellipse];
+  0;  1;  2;  3;  4;  5;  6;
+  0 -> 1;
+  0 -> 4;
+  0 -> 2;
+  0 -> 5;
+  1 -> 3;
+  1 -> 2;
+  5 -> 6;
+
+}
+
+
+****
+PDAG
+****
+digraph "no_name" {
+node [shape = ellipse];
+  0;  1;  2;  3;  4;  5;  6;
+  0 -> 4 [dir=none];
+  0 -> 2 [dir=none];
+  1 -> 3 [dir=none];
+  1 -> 2 [dir=none];
+  1 -> 0;
+  5 -> 6 [dir=none];
+  5 -> 0;
+
+}
+
+
+****
+Named DAG
+****
+[A,B,C,D,E,F,G]
+[G->F,F->A,D->B,B->A,B->C,A->E,A->C]
+
+digraph {
+    "G"->"F"
+    "F"->"A"
+    "D"->"B"
+    "B"->"A"
+    "B"->"C"
+    "A"->"E"
+    "A"->"C"
+    "C"
+    "E"
+}
