--- python/src/ContinuousMIIC2.i
@@ -1,10 +1,0 @@
-// SWIG file ContinuousMIIC2.i
-
-%{
-#include "otagrum/ContinuousMIIC2.hxx"
-%}
-
-%include ContinuousMIIC2_doc.i
-
-%copyctor OTAGRUM::ContinuousMIIC2;
-%include otagrum/ContinuousMIIC2.hxx
