--- python/src/ContinuousPC2.i
@@ -1,10 +1,0 @@
-// SWIG file ContinuousPC2.i
-
-%{
-#include "otagrum/ContinuousPC2.hxx"
-%}
-
-%include ContinuousPC2_doc.i
-
-%copyctor OTAGRUM::ContinuousPC2;
-%include "otagrum/ContinuousPC2.hxx"
