+++ python/test/CMakeLists.txt
@@ -65,9 +65,7 @@ endmacro ()
 ot_pyinstallcheck_test (Utils_std)
 ot_pyinstallcheck_test (docstring)
 ot_pyinstallcheck_test (ContinuousPC_std)
-ot_pyinstallcheck_test (ContinuousPC2_std)
 ot_pyinstallcheck_test (ContinuousMIIC_std)
-ot_pyinstallcheck_test (ContinuousMIIC2_std)
 ot_pyinstallcheck_test (TabuList_std)
 ot_pyinstallcheck_test (NamedJunctionTree_std)
 ot_pyinstallcheck_test (JunctionTreeBernsteinCopula_std)
