+++ python/test/t_ContinuousMIIC_std.py
@@ -44,5 +44,63 @@ def testAsiaDirichlet():
     sys.stdout.flush()
 
 
+def testMathis():
+    # extreme case to show that cmiic apply meek rule 3 correctly
+    import pyagrum as gum
+
+    dag = gum.DAG()
+    dag.addNodes(4)
+    for t, h in [(0, 1), (2, 1), (2, 3)]:
+        dag.addArc(t, h)
+
+    var_names = [f"X{i}" for i in range(4)]
+    named_dag = otagrum.NamedDAG(dag, var_names)
+
+    marginals = [ot.Uniform(0.0, 1.0) for _ in range(4)]
+
+    rho = 0.8
+    local_copulas = []
+    for i in range(4):
+        n_parents = named_dag.getParents(i).getSize()
+        dim = n_parents + 1
+        R = ot.CorrelationMatrix(dim)
+        for j in range(dim):
+            for k in range(j):
+                R[j, k] = rho
+        local_copulas.append(ot.NormalCopula(R))
+
+    cbn = otagrum.ContinuousBayesianNetwork(named_dag, marginals, local_copulas)
+
+    ot.RandomGenerator.SetSeed(42)
+    sample = cbn.getSample(100)
+
+    bn = gum.BayesNet()
+    for nid in dag.nodes():
+        bn.add(gum.LabelizedVariable(f"X{nid}", f"X{nid}", 2))
+    for nid in dag.nodes():
+        for cid in dag.children(nid):
+            bn.addArc(nid, cid)
+
+    learner = otagrum.ContinuousMIIC(sample)
+    learner.setAlpha(0.01)
+    learner.setVerbosity(False)
+
+    learned_pdag = learner.learnPDAG()
+    print("pdag:", learned_pdag)
+
+    learned_named_dag = learner.learnDAG()
+    learned_dag = learned_named_dag.getDAG()
+
+    bn2 = gum.BayesNet()
+    for nid in learned_dag.nodes():
+        bn2.add(gum.LabelizedVariable(f"X{nid}", f"X{nid}", 2))
+    for t, h in learned_dag.arcs():
+        bn2.addArc(t, h)
+
+    learned_cpdag = gum.EssentialGraph(bn2).pdag()
+    print("cpdag:", learned_cpdag)
+
+
 testSpecificInstance()
 testAsiaDirichlet()
+testMathis()
